+++ lib/libc_library.dart
@@ -1,0 +1,7 @@
+library libc.libc_library;
+
+import "dart:io";
+import "package:binary_interop/binary_interop.dart";
+
+part "src/libc_library/libc_library.dart";
+part "src/libc_library/libc_binary.dart";