+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+- Added generator of the library `libc` 
+- Initial version of the library `libc` was generated
+
 ## 0.0.3
 
 - Added function prototypes in `stdio.h`